requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.2
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
emails==0.6
et_xmlfile==2.0.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
lxml==6.0.2
more-itertools==10.8.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
pillow==11.3.0
premailer==3.10.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
